build: add modularization to support mobile-clip and clip
