+++ app/build.gradle
@@ -72,6 +72,10 @@ dependencies {
 
     implementation "androidx.compose.material:material-icons-extended:$compose_ui_version"
     implementation platform('androidx.compose:compose-bom:2023.03.00')
+//    implementation project(':features:mobileclip')
+    implementation project(':features:clip')
+    implementation project(':base:clipcore')
+    implementation project(':base:core')
     androidTestImplementation platform('androidx.compose:compose-bom:2023.03.00')
 
     def m3_version = "1.2.0-alpha06"
