--- app/src/main/java/me/grey/picquery/common/CoroutineExt.kt
@@ -1,8 +1,0 @@
-package me.grey.picquery.common
-
-import kotlinx.coroutines.Dispatchers
-
-val mainDispatcher = Dispatchers.Main
-val defaultDispatcher = Dispatchers.Default
-val unconfinedDispatcher = Dispatchers.Unconfined
-val ioDispatcher = Dispatchers.IO
