+++ base/clipcore/src/main/java/com/grey/picquery/library/EmbeddingMaker.kt
@@ -1,0 +1,8 @@
+package com.grey.picquery.library
+
+import android.graphics.Bitmap
+import java.nio.FloatBuffer
+
+interface EmbeddingMaker {
+    suspend fun makeBatchEmbedding(input: List<Bitmap>): FloatBuffer
+}
