+++ base/clipcore/src/main/java/com/grey/picquery/library/ImageEncoder.kt
@@ -1,0 +1,8 @@
+package com.grey.picquery.library
+
+import android.graphics.Bitmap
+
+interface ImageEncoder {
+
+    suspend fun encodeBatch(bitmaps: List<Bitmap>): List<FloatArray>
+}
