+++ base/clipcore/src/main/java/com/grey/picquery/library/textencoder/InputMaker.kt
@@ -1,0 +1,5 @@
+package com.grey.picquery.library.textencoder
+
+interface TextInputMaker<T> {
+    fun makeInput(text: String): T
+}
