+++ base/clipcore/src/main/java/com/grey/picquery/library/textencoder/TextEncoder.kt
@@ -1,0 +1,5 @@
+package com.grey.picquery.library.textencoder
+
+interface TextEncoder {
+    fun encode(input: String): FloatArray
+}
