+++ base/core/src/main/java/com/grey/picQuery/core/CoroutineExt.kt
@@ -1,0 +1,8 @@
+package com.grey.picQuery.core
+
+import kotlinx.coroutines.Dispatchers
+
+val mainDispatcher = Dispatchers.Main
+val defaultDispatcher = Dispatchers.Default
+val unconfinedDispatcher = Dispatchers.Unconfined
+val ioDispatcher = Dispatchers.IO
